Reorganize remaining RiveScript code into subpackage
