--- macros.go
@@ -1,8 +1,0 @@
-package rivescript
-
-// Interface for object macro handlers.
-
-type MacroInterface interface {
-	Load(name string, code []string)
-	Call(name string, fields []string) string
-}
