+++ rivescript/astmap.go
