+++ rivescript/inheritance.go
