+++ rivescript/utils.go
